+++ tests/test_io.py
@@ -7,6 +7,7 @@ from nnmnkwii.util import example_question_file
 import re
 from nose.tools import raises
 
+
 DATA_DIR = join(dirname(__file__), "data")
 
 
@@ -88,11 +89,11 @@ QS "L-Phone_Yuusei_Boin"           {*^a-*,*^i-*,*^u-*,*^e-*,*^o-*}
 CQS "e1" {/E:(\\NOTE)]}
     """
     binary_dict, continuous_dict = hts.load_question_set(
-        join(DATA_DIR, "test_jp_svs.hed"), append_hat_for_LL=False)
+        join(DATA_DIR, "test_jp_svs.hed"), append_hat_for_LL=False, convert_svs_pattern=True)
     input_phone_label = join(DATA_DIR, "song070_f00001_063.lab")
     labels = hts.load(input_phone_label)
     feats = fe.linguistic_features(labels, binary_dict, continuous_dict)
-    assert feats.shape == (74, 2)
+    assert feats.shape == (74, 3)
 
     # CQS e1: get the current MIDI number
     C_e1 = continuous_dict[0]
@@ -101,6 +102,19 @@ CQS "e1" {/E:(\\NOTE)]}
         if C_e1.search(context) is not None:
             from nnmnkwii.frontend import NOTE_MAPPING
             assert NOTE_MAPPING[C_e1.findall(context)[0]] == feats[idx, 1]
+
+    # CQS e57: get pitch diff
+    # In contrast to other continous features, the pitch diff has a prefix "m" or "p"
+    # to indiecate th sign of numbers.
+    C_e57 = continuous_dict[1]
+    for idx, lab in enumerate(labels):
+        context = lab[-1]
+        if "~p2+" in context:
+            assert C_e57.search(context).group(1) == "p2"
+            assert feats[idx, 2] == 2
+        if "~m2+" in context:
+            assert C_e57.search(context).group(1) == "m2"
+            assert feats[idx, 2] == -2
 
 
 def test_state_alignment_label_file():